backend/pyproject.toml
[project]
name = "jfxz-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "aiosqlite>=0.21.0",
  "argon2-cffi>=25.1.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "pydantic-settings>=2.8.1",
  "email-validator>=2.2.0",
  "sqlalchemy[asyncio]>=2.0.40",
  "agno[openai,postgres]>=2.6.3,<3",
  "apscheduler>=3.11.0",
  "typer>=0.15.0",
  "uvicorn>=0.34.0"
]

[project.scripts]
jfxz = "app.cli:cli"
jfxz-create-admin = "app.scripts.create_admin:main"

[project.optional-dependencies]
dev = [
  "coverage>=7.8.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "pytest-cov>=6.1.1",
  "pip-audit>=2.9.0",
  "ruff>=0.11.7"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["app"]
omit = ["app/__init__.py"]

[tool.coverage.report]
fail_under = 100
show_missing = true
skip_covered = true

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008", "E501"]

[tool.ruff.lint.per-file-ignores]
"app/scripts/create_admin.py" = ["E402"]
